# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API Server for all fests",
        "title": "FestAPI v2",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8000",
    "basePath": "/",
    "paths": {
...
        "/api/hospi/checkIn": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "If roomID is zero, visitor is not alloted room.",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Hospi"
                ],
                "summary": "Check in for the visitors.",
                "operationId": "CheckIn",
                "parameters": [
                    {
                        "description": "Check in request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CheckInRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Invalid Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/hospi/deleteRoom": {
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Delete a room with the provided ID.",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Hospi"
                ],
                "summary": "Delete a room.",
                "operationId": "DeleteRoom",
                "parameters": [
                    {
                        "description": "Delete room request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.DeleteRoomRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Invalid Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/hospi/getHostels": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Retrieve the details of the hostels.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Hospi"
                ],
                "summary": "Get all the Hostels",
                "operationId": "GetHostels",
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/dto.GetHostelsResponse"
                        }
                    },
                    "400": {
                        "description": "Hostels not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/hospi/getRooms": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Retrieve the details of the rooms along with the hostel name.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Hospi"
                ],
                "summary": "Get all the Rooms",
                "operationId": "GetRooms",
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/dto.GetRoomsResponse"
                        }
                    },
                    "400": {
                        "description": "Rooms not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/hospi/updateHostel": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Add/Update a new hostel with the provided details.",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Hospi"
                ],
                "summary": "Add/Update a new hostel.",
                "operationId": "AddUpdateHostel",
                "parameters": [
                    {
                        "description": "Add/update hostel request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.AddUpdateHostelRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Invalid Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/hospi/updateRoom": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Add/Update a new room with the provided details.",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Hospi"
                ],
                "summary": "Add/Update a new room.",
                "operationId": "AddUpdateRoom",
                "parameters": [
                    {
                        "description": "Add/update room request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.AddUpdateRoomRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Invalid Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/treasury/addBill": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Add a bill with its purpose: checkIn/checkOut/discount/fine/eventPass.",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Treasury"
                ],
                "summary": "Add a Bill.",
                "operationId": "AddBill",
                "parameters": [
                    {
                        "description": "Add bill request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.AddBillRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Invalid Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/treasury/townscript": {
            "post": {
                "description": "Make a payment using Townscript.",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Treasury"
                ],
                "summary": "Make a payment using Townscript.",
                "operationId": "Townscript",
                "parameters": [
                    {
                        "description": "Townscript request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.TownScriptRequest"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Secret Token",
                        "name": "secret",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Invalid Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.AddBillRequest": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "integer"
                },
                "mode": {
                    "type": "string"
                },
                "paid_to": {
                    "$ref": "#/definitions/models.AdminRole"
                },
                "ref_id": {
                    "type": "string"
                },
                "time": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "dto.AddUpdateHostelRequest": {
            "type": "object",
            "properties": {
                "gender": {
                    "$ref": "#/definitions/models.Gender"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "dto.AddUpdateRoomRequest": {
            "type": "object",
            "properties": {
                "capacity": {
                    "type": "integer"
                },
                "hostel_id": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "dto.CheckInRequest": {
            "type": "object",
            "properties": {
                "no_of_days": {
                    "type": "integer"
                },
                "room_id": {
                    "type": "integer"
                },
                "time": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
...
        "dto.DeleteRoomRequest": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                }
            }
        },
...
        "dto.GetHostelsResponse": {
            "type": "object",
            "properties": {
                "gender": {
                    "$ref": "#/definitions/models.Gender"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "dto.GetRoomsResponse": {
            "type": "object",
            "properties": {
                "capacity": {
                    "type": "integer"
                },
                "gender": {
                    "$ref": "#/definitions/models.Gender"
                },
                "hostel": {
                    "type": "string"
                },
                "hostelID": {
                    "type": "integer"
                },
                "room": {
                    "type": "string"
                },
                "roomID": {
                    "type": "integer"
                }
            }
        },
...
        "dto.TownScriptRequest": {
            "type": "object",
            "properties": {
                "answerList": {
                    "type": "array",
                    "items": {
                        "type": "object",
                        "properties": {
                            "answer": {
                                "type": "string"
                            },
                            "question": {
                                "type": "string"
                            },
                            "uniqueQuestionId": {
                                "type": "integer"
                            }
                        }
                    }
                },
                "currency": {
                    "type": "string"
                },
                "customQuestion1": {
                    "type": "string"
                },
                "customQuestion20": {
                    "type": "string"
                },
                "discountAmount": {
                    "type": "integer"
                },
                "discountCode": {
                    "type": "string"
                },
                "eventCode": {
                    "type": "string"
                },
                "eventName": {
                    "type": "string"
                },
                "registrationTimestamp": {
                    "type": "string"
                },
                "ticketName": {
                    "type": "string"
                },
                "ticketPrice": {
                    "type": "integer"
                },
                "uniqueOrderId": {
                    "type": "string"
                },
                "userEmailId": {
                    "type": "string"
                },
                "userName": {
                    "type": "string"
                }
            }
        },
...
        "models.AdminRole": {
            "type": "string",
            "enum": [
                "ADMIN",
                "OC",
                "PR",
                "TREZ",
                "CORE"
            ],
            "x-enum-varnames": [
                "ADMIN",
                "OC",
                "PR",
                "TREZ",
                "CORE"
            ]
        },
        "models.Gender": {
            "type": "string",
            "enum": [
                "MALE",
                "FEMALE",
                "OTHER"
            ],
            "x-enum-varnames": [
                "MALE",
                "FEMALE",
                "OTHER"
            ]
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "Authorization token",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}